Refactor: Tighten UserRepository by extracting health profile and leader parsing logic
Moves `parseLeadersJson` to `UserEntity` companion object.
Moves `getHealthProfile` and `updateUserHealthProfile` to `HealthRepositoryImpl`.
Removes `healthRepository` dependency from `UserRepositoryImpl`.
Fixes dependent tests for the constructor updates.

Rebased onto master and resolved the conflicts with the work that landed
since:

- UserRepository/UserRepositoryImpl: kept master's new `searchUsers` and
  DAO-backed `saveUser`/`getPendingSyncUsers`, dropped only
  `parseLeadersJson`, and removed the imports left orphaned by the move.
- HealthViewModel: kept master's parallel `async` loads, sourcing the
  health profile from `healthRepository`.
- VoicesAdapter: master now takes `leadersList` as a constructor
  parameter, so the adapter no longer parses leaders itself and the moved
  helper is not needed there.
- Updated the leader-parsing call sites master added after this branch
  (`VoicesFragment`, `TeamsVoicesFragment`, `ReplyActivity`) to
  `UserEntity.parseLeadersJson`, and dropped the `UserRepository`
  dependencies that became unused in `ReplyActivity` and
  `LeadersViewModel`.
- `HealthRepositoryImpl` referenced an undeclared `userDao`; injected
  `UserDao` and updated `HealthRepositoryImplTest` accordingly.
