+++ app/src/main/java/org/ole/planet/myplanet/repository/UserRepository.kt
@@ -7,7 +7,6 @@ import org.ole.planet.myplanet.model.Achievement
 import org.ole.planet.myplanet.model.AchievementData
 import org.ole.planet.myplanet.model.DashboardProfile
 import org.ole.planet.myplanet.model.MemberInfo
-import org.ole.planet.myplanet.model.MyHealth
 import org.ole.planet.myplanet.model.User
 import org.ole.planet.myplanet.model.UserEntity
 
@@ -16,8 +15,6 @@ interface UserRepository {
     suspend fun getSavedUsers(): List<User>
     suspend fun upsertSavedUser(name: String?, encryptedPassword: String?, source: String, userProfile: String?, userName: String?)
     suspend fun resetGuestAsMember(username: String?)
-    suspend fun getHealthProfile(userId: String): MyHealth?
-    suspend fun updateUserHealthProfile(userId: String, userData: Map<String, Any?>)
 
     suspend fun getUserById(userId: String): UserEntity?
     suspend fun getDashboardProfile(userId: String): DashboardProfile
@@ -34,7 +31,6 @@ interface UserRepository {
     suspend fun getPendingSyncUsers(limit: Int): List<UserEntity>
     suspend fun searchUsers(query: String, sortField: String, descending: Boolean): List<UserEntity>
     suspend fun saveUser(user: UserEntity)
-    fun parseLeadersJson(jsonString: String): List<UserEntity>
     suspend fun ensureUserSecurityKeys(userId: String): UserEntity?
     suspend fun fetchUserSecurityData(name: String)
     suspend fun updateSecurityData(
