+++ app/src/test/java/org/ole/planet/myplanet/repository/UserRepositoryImplTest.kt
@@ -106,7 +106,6 @@ class UserRepositoryImplTest {
             mockk(relaxed = true),
             mockk(relaxed = true),
             mockk(relaxed = true),
-            mockk(relaxed = true),
             userDao,
             mockk(relaxed = true)
         )
